Move rails loading to just before delayed
