+++ .prettierignore
@@ -2,3 +2,4 @@ prism.js
 blocks/**/**.css
 styles/**/**.css
 scripts/prism-grammars/**
+scripts/**/**.css
