+++ scripts/rails/rails.scss
@@ -1,4 +1,4 @@
-@import './sass/breakpoints';
+@import '../../styles/sass/breakpoints';
 
 .docs {
   .rail-toggle {
